fd and null pointer protections, return (-1) if error occurs

+++ ft_putchar.c
@@ -6,13 +6,13 @@
 /*   By: jfarinha <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 11:48:00 by jfarinha          #+#    #+#             */
-/*   Updated: 2018/04/17 15:35:10 by jfarinha         ###   ########.fr       */
+/*   Updated: 2018/04/18 07:29:26 by jfarinha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_putchar(int c)
+int		ft_putchar(int c)
 {
 	return (ft_putchar_fd(c, STDOUT));
 }
