+++ ft_putendl.c
@@ -6,13 +6,13 @@
 /*   By: jfarinha <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 11:50:23 by jfarinha          #+#    #+#             */
-/*   Updated: 2018/04/17 15:37:31 by jfarinha         ###   ########.fr       */
+/*   Updated: 2018/04/18 07:36:38 by jfarinha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_putendl(const char *str)
+int		ft_putendl(const char *str)
 {
 	return (ft_putendl_fd(str, STDOUT));
 }
